requirements.txt
Flask==3.0.3
google-genai==1.7.0
google-cloud-texttospeech==2.17.2
gunicorn==22.0.0